requirements.txt
dtaidistance
fastparquet
kneed 
matplotlib 
numpy 
pandas 
pandas-ta
pyarrow
python-dotenv
PyWavelets
quantstats 
scikit-learn
sktime
tslearn
seaborn
sphinx
sphinx_rtd_theme
sphinx_autodoc_typehints
